refactor: playground env select
